move convert event to project detail screen
