+++ USFMConverter/UI/Pages/ProjectDetailScreen.axaml.cs
@@ -55,6 +55,18 @@ namespace USFMConverter.UI.Pages
             }
         }
 
+        public event EventHandler<RoutedEventArgs> ConvertStart
+        {
+            add
+            {
+                AddHandler(StartConvertEvent, value);
+            }
+            remove
+            {
+                RemoveHandler(StartConvertEvent, value);
+            }
+        }
+
         public static readonly RoutedEvent<RoutedEventArgs> ShowBackgroundOverlayEvent =
             RoutedEvent.Register<FileView, RoutedEventArgs>(
                 nameof(ShowBackgroundOverlay),
@@ -70,6 +82,12 @@ namespace USFMConverter.UI.Pages
         public static readonly RoutedEvent<RoutedEventArgs> StartNewProjectEvent =
             RoutedEvent.Register<FileView, RoutedEventArgs>(
                 nameof(StartNewProject),
+                RoutingStrategies.Bubble
+            );
+
+        public static readonly RoutedEvent<RoutedEventArgs> StartConvertEvent =
+            RoutedEvent.Register<FileView, RoutedEventArgs>(
+                nameof(ConvertStart),
                 RoutingStrategies.Bubble
             );
